+++ tests/testthat/test-tdigest.R
@@ -4,7 +4,14 @@ td <- td_create(10)
 
 expect_is(td, "tdigest")
 
+expect_true(is_tdigest(td))
 expect_equal(td_total_count(td), 0)
+
+expect_true(is.nan(td_value_at(td, 0)))
+expect_true(is.nan(td_value_at(td, 0.5)))
+expect_true(is.nan(td_value_at(td, 1)))
+expect_true(is.nan(td_value_at(td, -0.1)))
+expect_true(is.nan(td_value_at(td, 1.1)))
 
 td_add(td, 0, 1)
 td_add(td, 10, 1)
@@ -14,12 +21,23 @@ expect_equal(td_total_count(td), 2)
 expect_equal(td_value_at(td, 0.1), 0)
 expect_equal(td_value_at(td, 0.5), 5)
 
+td <- td_create(1000)
+td_add(td, 1, 1)
+td_add(td, 10, 1)
 
-context("bigger test")
+expect_equal(td_quantile_of(td, 0.99), 0)
+expect_equal(td_quantile_of(td, 1), 0.25)
+expect_equal(td_quantile_of(td, 5.5), 0.5)
+
+context("bigger, vectorised test")
 
 set.seed(1492)
 x <- sample(0:100, 1000000, replace = TRUE)
 td <- tdigest(x, 1000)
+
+expect_true(is_tdigest(td))
+expect_false(is_tdigest(x))
+
 expect_equal(td_total_count(td), 1000000)
 
 expect_identical(
@@ -27,4 +45,9 @@ expect_identical(
     tquantile(td, c(0, .01, .1, .2, .3, .4, .5, .6, .7, .8, .9, .99, 1))
   ),
   c(0L, 0L, 9L, 19L, 29L, 39L, 50L, 60L, 70L, 80L, 90L, 99L, 100L)
+)
+
+expect_identical(
+  as.integer(quantile(td)),
+  c(0L, 24L, 50L, 75L, 100L)
 )