+++ NEWS.md
@@ -1,2 +1,8 @@
+0.2.0
+* Added input vaildity checks
+* Added `quantile()` function S3 implementation for `tdigest` objects
+* Added examples
+* Added more tests
+
 0.1.0 
 * Initial release